+++ RedPandaIDE/widgets/aboutdialog.cpp
@@ -1,6 +1,8 @@
 #include "aboutdialog.h"
 #include "ui_aboutdialog.h"
 #include "../systemconsts.h"
+#include "../version.h"
+
 
 AboutDialog::AboutDialog(QWidget *parent) :
     QDialog(parent),
@@ -8,9 +10,20 @@ AboutDialog::AboutDialog(QWidget *parent) :
 {
     ui->setupUi(this);
     ui->lblTitle->setText(ui->lblTitle->text() + tr("Version: ") + DEVCPP_VERSION);
+
+#ifdef  __GNUC__
     ui->lblContent->setText(ui->lblContent->text()
                             .arg(qVersion())
-                            .arg("GCC 10.3.0",__DATE__, __TIME__));
+                            .arg(QString("GCC %1.%2")
+                                 .arg(__GNUC__)
+                                 .arg(__GNUC_MINOR__)
+                                 ,__DATE__, __TIME__));
+#else
+    ui->lblContent->setText(ui->lblContent->text()
+                            .arg(qVersion())
+                            .arg("Non-GCC Compiler"
+                                 ,__DATE__, __TIME__));
+#endif
 }
 
 AboutDialog::~AboutDialog()